[basic-example][m] - remove everything related to multiple pages
